+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Add in-game menu (graceful way to quit the game room)
 - Fix rare font atlas artifacts
+- Change settings button in Main Menu
+- Add settings for Practice Mode 
 
 ## 1.0.12